@@ -1,11 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Fluent.Net.SyntaxTest\Fluent.Net.SyntaxTest.csproj" />
+    <ProjectReference Include="../Fluent.Net.SyntaxTest/Fluent.Net.SyntaxTest.csproj" />
   </ItemGroup>
 
   <ItemGroup>
